Move notes_editor module into fields dir
